some minor refactor around request path
