+++ src/AutoRest.CSharp/Mgmt/Models/RequestPath.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Collections;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
@@ -24,10 +25,17 @@ namespace AutoRest.CSharp.Mgmt.Models
         private const string _providerPath = "/subscriptions/{subscriptionId}/providers/{resourceProviderNamespace}";
         private const string _featurePath = "/subscriptions/{subscriptionId}/providers/Microsoft.Features/providers/{resourceProviderNamespace}/features";
 
-        internal const string ManagementGroupScopePrefix = "/providers/Microsoft.Management/managementGroups";
-        internal const string ResourceGroupScopePrefix = "/subscriptions/{subscriptionId}/resourceGroups";
-        internal const string SubscriptionScopePrefix = "/subscriptions";
-        internal const string TenantScopePrefix = "/tenants";
+        private const string ManagementGroupScopePrefix = "/providers/Microsoft.Management/managementGroups";
+        private const string ResourceGroupScopePrefix = "/subscriptions/{subscriptionId}/resourceGroups";
+        private const string SubscriptionScopePrefix = "/subscriptions";
+        private const string TenantScopePrefix = "/tenants";
+
+        public static readonly IReadOnlyList<RequestPath> KnownParentResourcePaths = new[]
+        {
+            ResourceGroup,
+            Subscription,
+            ManagementGroup
+        };
 
         /// <summary>
         /// This is a placeholder of request path for "any" resources in other RPs
@@ -194,6 +202,37 @@ namespace AutoRest.CSharp.Mgmt.Models
                 return true;
             }
             return false;
+        }
+
+        private static ConcurrentDictionary<RequestPath, RequestPath> _scopePathCache = new ConcurrentDictionary<RequestPath, RequestPath>();
+
+        public RequestPath GetScopePath()
+        {
+            if (_scopePathCache.TryGetValue(this, out var result))
+                return result;
+
+            result = CalculateScopePath(this);
+            _scopePathCache.TryAdd(this, result);
+            return result;
+        }
+
+        private static RequestPath CalculateScopePath(RequestPath requestPath)
+        {
+            var indexOfProvider = requestPath.ToList().LastIndexOf(Segment.Providers);
+            // if there is no providers segment, myself should be a scope request path. Just return myself
+            if (indexOfProvider >= 0)
+            {
+                if (indexOfProvider == 0 && requestPath.SerializedPath.StartsWith(RequestPath.ManagementGroupScopePrefix, StringComparison.InvariantCultureIgnoreCase))
+                    return RequestPath.ManagementGroup;
+                return new RequestPath(requestPath.Take(indexOfProvider));
+            }
+            if (requestPath.SerializedPath.StartsWith(RequestPath.ResourceGroupScopePrefix, StringComparison.InvariantCultureIgnoreCase))
+                return RequestPath.ResourceGroup;
+            if (requestPath.SerializedPath.StartsWith(RequestPath.SubscriptionScopePrefix, StringComparison.InvariantCultureIgnoreCase))
+                return RequestPath.Subscription;
+            if (requestPath.SerializedPath.Equals(RequestPath.TenantScopePrefix))
+                return RequestPath.Tenant;
+            return requestPath;
         }
 
         /// <summary>
